Rental Properties has foreign key to personal info rather than rentalUser

+++ home-rental-app/src/main/java/com/rental/repo/RentalPropsRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.rental.entity.RentalProps;
 
-public interface RentalPropsRepo extends JpaRepository<RentalProps, Integer> {
+public interface RentalPropsRepo extends JpaRepository<RentalProps, Long> {
 
 }
